Align renderscript Allocation memory for AVX
The Allocation class aligned to 16 byte. This breaks
some AVX instructions that require 32 byte alignment.

This patch adds support for 32 byte alignment.
For efficiency it switches from the default alignment
to the natural alignment based on the element type and
rounds it to be a multiple of sizeof(void*) and a power
of two.

Author: Andy Ross <andy.ross@windriver.com>
Author: Michael Liao <michael.liao@intel.com>

Issue: AXIA-1222
Category: AOSP Improvement
Upstream Candidate: Yes
Domain: AOSP-frameworks

Change-Id: I23089ed868839b12bb2784249cb5998138012935